Run migrations before `storageCache` gets constructed
Otherwise, the migrations will not be maintainable.
For example, the favorites migration used `storageCache.setItem` to store the favorites in the new format.
However, if we decide to change the favorite representation again in the future, this code will become invalid.
