+++ src/services/storageUtils.ts
@@ -5,7 +5,7 @@ import { ExerciseKey, Favorite, FIRST_EXERCISE_FOR_REPETITION, VOCABULARY_ITEM_T
 import { UserVocabularyItem, VocabularyItem } from '../constants/endpoints'
 import { VocabularyItemResult } from '../navigation/NavigationTypes'
 import { RepetitionService } from './RepetitionService'
-import { getStorageItemOr, STORAGE_VERSION, StorageCache } from './Storage'
+import { getStorageItem, getStorageItemOr, STORAGE_VERSION, StorageCache, storageKeys } from './Storage'
 import { calculateScore, vocabularyItemToFavorite } from './helpers'
 
 export const FAVORITES_KEY_VERSION_0 = 'favorites'
@@ -71,31 +71,48 @@ export const saveExerciseProgress = async (
 const compareFavorites = (favorite1: Favorite, favorite2: Favorite) =>
   favorite1.id === favorite2.id && favorite1.vocabularyItemType === favorite2.vocabularyItemType
 
-export const migrateToNewFavoriteFormat = async (storageCache: StorageCache): Promise<void> => {
-  const parsedVocabularyItems = await getStorageItemOr<number[]>(FAVORITES_KEY_VERSION_0, [])
-  if (parsedVocabularyItems.length === 0) {
+export const migrate0To1 = async (): Promise<void> => {
+  const parsedFavorites = await getStorageItemOr<number[]>(FAVORITES_KEY_VERSION_0, [])
+  if (parsedFavorites.length === 0) {
     return
   }
-  await storageCache.setItem(
-    'favorites',
-    parsedVocabularyItems.map((item: number) => ({
-      id: item,
-      vocabularyItemType: VOCABULARY_ITEM_TYPES.lunesStandard,
-    })),
+  await AsyncStorage.setItem(
+    'favorites-2',
+    JSON.stringify(
+      parsedFavorites.map((item: number) => ({
+        id: item,
+        vocabularyItemType: 'lunes-standard',
+      })),
+    ),
   )
   await AsyncStorage.removeItem(FAVORITES_KEY_VERSION_0)
 }
 
-export const migrateStorage = async (storageCache: StorageCache): Promise<void> => {
-  const lastVersion = storageCache.getItem('version')
+export const migrateStorage = async (): Promise<void> => {
+  const getStorageVersion = async (): Promise<number> => {
+    const version = await getStorageItemOr<number | null>(storageKeys.version, null)
+    if (version !== null) {
+      return version
+    }
+
+    // If there is no version number stored yet,
+    // this is either a new installation or an update from a version where this field did not exist yet.
+    // In the former case, the storage version should be the latest version to avoid unnecessary startup work.
+    // In the latter case, we should use 0 as the version number so that all migrations are run.
+    // To differentiate between the two cases, we can use the fact that `selectedProfessions` is null if and only if the startup screen was not completed yet.
+    const selectedProfessions = await getStorageItem('selectedProfessions')
+    return selectedProfessions === null ? STORAGE_VERSION : 0
+  }
+
+  const lastVersion = await getStorageVersion()
   switch (lastVersion) {
     case 0:
-      await migrateToNewFavoriteFormat(storageCache)
+      await migrate0To1()
       break
   }
 
   if (lastVersion !== STORAGE_VERSION) {
-    await storageCache.setItem('version', STORAGE_VERSION)
+    await AsyncStorage.setItem(storageKeys.version, STORAGE_VERSION.toString())
   }
 }
